apps/knowledge/evals/pyproject.toml
[project]
name = "knowledge-evals"
version = "0.1.0"
description = "DeepEval quality evaluations for knowledge articles"
requires-python = ">=3.12"
dependencies = [
    "deepeval>=2.5.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "langgraph>=0.4.0",
    "langchain-openai>=0.3.0",
    "langchain-core>=0.3.0",
    "langgraph-checkpoint-postgres>=2.0.0",
    "sentence-transformers>=3.0.0",
    "chromadb>=1.0.0",
    "langchain-text-splitters>=0.3.0",
    "langchain-community>=0.3.0",
    "deepteam>=1.0",
    "psycopg2-binary>=2.9.0",
    "pgvector>=0.3.0",
    "fastembed>=0.4.0",
]

[tool.pytest.ini_options]
env = [
    "DEEPEVAL_TELEMETRY_OPT_OUT=YES",
]
